requirements.txt
# Universal FRA/SFRA Parser Dependencies
# SIH 2025 PS 25190

# PRODUCTION DEPENDENCIES
# ======================

# Core data processing (allow patch updates for security fixes)
pandas>=2.0.0,<3.0.0  # Updated to stable 2.x series, allow patches
numpy>=1.24.0,<2.0.0  # Updated for security and performance, allow patches

# Simulator dependencies
scipy>=1.10.0,<2.0.0  # For signal processing and transfer function computation
matplotlib>=3.7.0,<4.0.0  # For Bode plot visualization

# AI/ML Ensemble dependencies
torch>=2.0.0,<3.0.0  # PyTorch for deep learning (1D CNN, ResNet18) - ARM compatible
torchvision>=0.15.0,<1.0.0  # For pre-trained models and transforms
scikit-learn>=1.3.0,<2.0.0  # For One-Class SVM and metrics
Pillow>=10.0.1,<11.0.0  # SECURITY FIX: Updated from 8.x (CVE fixes)

# Web Dashboard dependencies
streamlit>=1.28.0,<2.0.0  # Interactive web application framework
plotly>=5.17.0,<6.0.0  # Interactive visualizations (Bode plots, bar charts)
seaborn>=0.12.0,<1.0.0  # Statistical data visualization (heatmaps)

# PDF Report Generation
reportlab>=4.0.0,<5.0.0  # IEC-compliant PDF report generation

# Edge Deployment (ONNX)
onnx>=1.14.0,<2.0.0  # ONNX model format
onnxruntime>=1.15.0,<2.0.0  # ONNX Runtime for edge inference (CPU optimized)

# Security and XML parsing
defusedxml>=0.7.1,<1.0.0  # SECURITY: Safe XML parsing (prevents XXE attacks)
lxml>=4.9.3,<5.0.0  # Robust XML parsing with security patches

# Configuration management
python-dotenv>=1.0.0,<2.0.0  # Environment variable management

# Progress tracking
tqdm>=4.66.0,<5.0.0  # Progress bars for long operations


# DEVELOPMENT DEPENDENCIES
# =========================
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.4.0,<8.0.0  # Testing framework
pytest-cov>=4.1.0,<5.0.0  # Coverage reporting
pytest-mock>=3.11.0,<4.0.0  # Mocking for tests
hypothesis>=6.88.0,<7.0.0  # Property-based testing

# Code quality
black>=23.9.0,<24.0.0  # Code formatter
flake8>=6.1.0,<7.0.0  # Linting
mypy>=1.5.0,<2.0.0  # Type checking
isort>=5.12.0,<6.0.0  # Import sorting

# Security scanning
bandit>=1.7.5,<2.0.0  # Security issue scanner
safety>=2.3.5,<3.0.0  # Dependency vulnerability scanner
pip-audit>=2.6.0,<3.0.0  # Audit pip packages for known vulnerabilities

# Documentation
sphinx>=7.2.0,<8.0.0  # Documentation generator
sphinx-rtd-theme>=1.3.0,<2.0.0  # ReadTheDocs theme

# Performance profiling
memory-profiler>=0.61.0,<1.0.0  # Memory usage profiling
line-profiler>=4.1.0,<5.0.0  # Line-by-line profiling


# NOTES
# =====
# - Versions allow patch updates for security fixes (e.g., 2.0.x series)
# - Major version locked to prevent breaking changes (e.g., <3.0.0)
# - Security vulnerabilities addressed (Pillow 8.x -> 10.x, added defusedxml)
# - Added development tools for code quality and security
# - Use 'pip freeze > requirements.lock' to generate exact version lock file
# - For production: pip install -r requirements.txt
# - For development: pip install -r requirements.txt -r requirements-dev.txt
#
# Standard library modules (included with Python, no installation needed):
# - xml.etree.ElementTree (replaced by defusedxml for security)
# - os, csv, datetime, typing, warnings, unittest, tempfile, shutil
# - io, pickle, argparse, json, time, logging, gc, sys

# GPU support optional (CUDA) - CPU fallback available
# Web dashboard runs locally on localhost
# Edge deployment optimized for Raspberry Pi (ARM)
# No external API dependencies - fully offline capable
